+++ CHANGELOG.md
@@ -3,12 +3,16 @@
 
 When stuff changes, it's described here.
 
-## 2015-04-16
+## 2015-04-27 – v0.3.3
+ * Updated JSONP object serializer to handle arrays slightly better
+ * Added support for Ad product pools, searches the global pool by default
+
+## 2015-04-16 – v0.3.2
  * Added key parameter to ad and product widgets
  * Fixed: when automatically inserting a product widget into an ad, use the ad's key instead of assuming a global key is defined
  * Fixed: call `findWidgetKey` after `parseConfiguration` in the widget init process, so the given key data element can be used
 
-## 2015-04-14
+## 2015-04-14 – v0.3.1
  * Added new and improved version of the initial AdCommerce widget, the Ad Widget
  * Added and integrated modified NanoModal library
  * Added Array.every polyfill because IE sucks